some modification and addw

+++ libft/char_conv1.c
@@ -98,7 +98,7 @@ char	*wchart_conv(unsigned int c, t_conv *data, int *ln)
 		ft_intfjoin(c, &converted, 0);
 	else
 		len = get_unicode_len(c);
-	if ((unsigned int)len > MB_CUR_MAX || len == -1)
+	if ((unsigned int)len > (unsigned int)MB_CUR_MAX || len == -1)
 	{
 		ft_strdel(&converted);
 		return (NULL);
